Several changes to support M1 Macbook Pro
The trickiest thing was getting the paths to work correctly. MacOS's
`path_helper` was overriding the paths I was setting and was prepending
e.g. `/usr/bin` in front of homebrew paths, which was causing things
like `vim` to load the old MacOS version. I fixed this by tweaking the
path logic to always prepend and moving that into `.zshrc`.